module update finished

+++ scheduler/src/main/java/com/charles/spider/scheduler/module/ComponentProxy.java
@@ -1,11 +1,10 @@
-package com.charles.spider.scheduler.moudle;
+package com.charles.spider.scheduler.module;
 
-
-import com.charles.spider.scheduler.persist.Service;
-import com.charles.spider.transfer.entity.ModuleType;
-import com.charles.spider.transfer.entity.Module;
 import com.charles.spider.query.Query;
 import com.charles.spider.query.condition.Condition;
+import com.charles.spider.scheduler.persist.Service;
+import com.charles.spider.transfer.entity.Module;
+import com.charles.spider.transfer.entity.ModuleType;
 import org.apache.commons.codec.digest.DigestUtils;
 import org.apache.commons.io.FileUtils;
 
@@ -14,46 +13,29 @@ import java.nio.file.*;
 import java.util.Date;
 import java.util.List;
 
-/**
- * Created by lq on 17-3-16.
- */
-public class ModuleAgent {
+public class ComponentProxy {
 
-    private Path base = null;
+    private Path base;
+    private Service<Module> service;
+    private ModuleType componentType;
 
-    private ModuleType type = null;
+    public ComponentProxy(ModuleType type, Service<Module> service, Path basePath) {
+        this.base = basePath;
 
-    private  Service<Module> service;
+        this.componentType = type;
 
-
-    public ModuleAgent(ModuleType type, Path path, Service<Module> service) {
-        this.type = type;
-        this.base = path;
         this.service = service;
     }
 
-
-    protected Path base() {
-        return base;
-    }
-
-
-    protected Service<Module> service() {
-        return this.service;
-    }
-
-
-    public ModuleType type() {
-        return this.type;
-    }
+    protected Service<Module> service(){return service;}
 
 
     public Module get(String name) throws IOException {
-        Path path = Paths.get(base().toString(), name);
+
+        Path path = Paths.get(base.toString(), name);
         if (!Files.exists(path) || !Files.isDirectory(path))
             return null;
-
-        Module module = service().single(Query.Condition(Condition.where("name").is(name)));
+        Module module = service.single(Query.Condition(Condition.where("name").is(name)));
         if (module == null) return null;
         if (module.getState() == Module.State.TMP) {
             Path old = Paths.get(path.toString(), "data");
@@ -61,7 +43,7 @@ public class ModuleAgent {
             Files.copy(tmp, old, StandardCopyOption.REPLACE_EXISTING);
 
             module.setState(Module.State.VALID);
-            int count = service().update(module,
+            int count = service.update(module,
                     Condition.where("id").is(module.getId())
                             .and(Condition.where("update_time").is(module.getUpdateTime())));
             Files.delete(tmp);
@@ -72,7 +54,7 @@ public class ModuleAgent {
 
 
     public Module save(byte[] data, String name, ModuleType type, String description, boolean override) throws Exception {
-        Path path = Paths.get(base().toString(), name);
+        Path path = Paths.get(base.toString(), name);
         Path old = Paths.get(path.toString(), "data");
         Path tmp = Paths.get(path.toString(), "data.tmp");
 
@@ -84,34 +66,28 @@ public class ModuleAgent {
 
         String hash = DigestUtils.sha1Hex(data);
 
-        Module module = service().single(Query.Condition(Condition.where("name").is(name)));
+        Module module = service.single(Query.Condition(Condition.where("name").is(name)));
         if (module == null) {
-            module = new Module();
-            module.setPath(path.toString());
-            module.setName(name);
-            module.setDetail(description);
-            module.setType(type);
-            module.setUpdateTime(new Date());
-            module.setHash(hash);
-            module = service().insert(module);
+            module = buildModule(name, type, path.toString(), hash, description);
+            module = service.insert(module);
         }
 
         if (module.getState() == Module.State.VALID && module.getHash().equals(hash))
-            throw new ModuleNoChangeException();
+            throw new ComponentNoChangeException();
 
         module.setHash(hash);
 
         //写入临时文件
         Files.write(tmp, data);
         module.setState(Module.State.TMP);
-        service().update(module,
+        service.update(module,
                 Condition.where("id").is(module.getId())
                         .and(Condition.where("update_time").is(module.getUpdateTime())));
 
         Files.copy(tmp, old, StandardCopyOption.REPLACE_EXISTING);
 
         module.setState(Module.State.VALID);
-        int count = service().update(module,
+        int count = service.update(module,
                 Condition.where("id").is(module.getId())
                         .and(Condition.where("update_time").is(module.getUpdateTime())));
 
@@ -122,36 +98,38 @@ public class ModuleAgent {
         return module;
     }
 
-
     public List<Module> select(Query query) {
         if (query == null) query = new Query();
 
-        query.addCondition(Condition.where("type").is(type()));
+        query.addCondition(Condition.where("type").is(componentType));
 
 
-        return service().select(query);
+        return service.select(query);
     }
 
 
-    public void delete(Query query) throws IOException {
-        Module module = service().single(query);
+    public Module delete(Query query) throws IOException {
+        Module module = service.single(query);
 
-        if (module.getState() != Module.State.NULL) {
-            int count = service().delete(query);
+        if (module!=null&& module.getState() != Module.State.NULL) {
+            int count = service.delete(query);
 
             if (count == 1) {
-                Path path = Paths.get(base().toString(), module.getName());
+                Path path = Paths.get(base.toString(), module.getName());
                 FileUtils.deleteDirectory(path.toFile());
             }
-
-
         }
+        return module;
     }
 
-
-    public Object object(String moduleName, String className) throws IOException, ModuleBuildException {
-        return null;
+    protected Module buildModule(String name, ModuleType type, String path, String hash, String desc) {
+        Module module = new Module();
+        module.setName(name);
+        module.setPath(path);
+        module.setDetail(desc);
+        module.setType(type);
+        module.setUpdateTime(new Date());
+        module.setHash(hash);
+        return module;
     }
-
-
 }
